+++ cli/src/tap/mod.rs
@@ -31,7 +31,7 @@ use {
         loading,
         notify_success,
         prelude::*,
-        sui::get_nexus_client,
+        sui::{get_nexus_client, get_read_only_nexus_client},
     },
     convert_case::{Case, Casing},
     nexus_sdk::{