+++ cli/src/tool/tool_set_invocation_cost.rs
@@ -14,7 +14,7 @@ pub(crate) async fn set_tool_invocation_cost(
     command_title!("Setting '{invocation_cost}' invocation cost for tool '{tool_fqn}'");
 
     let nexus_client = get_nexus_client(sui_gas_coin, sui_gas_budget).await?;
-    let signer = nexus_client.signer();
+    let signer = nexus_client.signer().map_err(NexusCliError::Nexus)?;
     let address = signer.get_active_address();
     let nexus_objects = &*nexus_client.get_nexus_objects();
     let crawler = nexus_client.crawler();