+++ sdk/src/types/nexus_objects.rs
@@ -7,6 +7,8 @@ use crate::move_bindings::{
     scheduler::task as scheduler_task_move,
     workflow::execution as execution_move,
 };
+#[cfg(feature = "nexus")]
+use std::sync::Arc;
 use {
     crate::{
         move_bindings::{
@@ -26,8 +28,6 @@ use {
     serde::{Deserialize, Serialize},
     sui_move::{MoveStruct, MoveType},
 };
-#[cfg(feature = "nexus")]
-use {std::sync::Arc, tokio::sync::Mutex};
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct UsTokenConfig {
@@ -179,7 +179,7 @@ impl NexusObjects {
     #[cfg(feature = "nexus")]
     pub async fn resolve_workflow_original_pkg_id(
         &mut self,
-        client: &Arc<Mutex<sui::grpc::Client>>,
+        client: &Arc<sui::grpc::Client>,
     ) -> anyhow::Result<()> {
         use sui::traits::FieldMaskUtil;
 
@@ -190,8 +190,8 @@ impl NexusObjects {
             .with_read_mask(field_mask);
 
         let response = client
-            .lock()
-            .await
+            .as_ref()
+            .clone()
             .ledger_client()
             .get_object(request)
             .await
@@ -230,7 +230,7 @@ impl NexusObjects {
     #[cfg(feature = "nexus")]
     pub async fn resolve_scheduler_original_pkg_id(
         &mut self,
-        client: &Arc<Mutex<sui::grpc::Client>>,
+        client: &Arc<sui::grpc::Client>,
     ) -> anyhow::Result<()> {
         use sui::traits::FieldMaskUtil;
 
@@ -241,8 +241,8 @@ impl NexusObjects {
             .with_read_mask(field_mask);
 
         let response = client
-            .lock()
-            .await
+            .as_ref()
+            .clone()
             .ledger_client()
             .get_object(request)
             .await